requirements.txt
PyYAML==5.4.1
jinja2
kubernetes
jc
